#36 - libcore: std feature
