--- src/libcore/features/std/std_.rs
@@ -1,74 +1,0 @@
-//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
-//  SPDX-License-Identifier: MIT
-
-//! std interface
-use crate::{
-    core::{
-        apply::Core as _,
-        exception::{self, Condition, Exception},
-        frame::Frame,
-        mu::Mu,
-        types::Type,
-    },
-    features::Features,
-    types::{
-        cons::{Cons, Core as _},
-        fixnum::Fixnum,
-        vector::{Core as _, Vector},
-    },
-};
-
-pub trait MuFunction {
-    fn std_command(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn std_exit(_: &Mu, fp: &mut Frame) -> exception::Result<()>;
-}
-
-impl MuFunction for Features {
-    fn std_command(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let command = fp.argv[0];
-        let args = fp.argv[1];
-
-        fp.value = match mu.fp_argv_check("spawn", &[Type::String, Type::List], fp) {
-            Ok(_) => {
-                let mut argv = vec![];
-
-                for cons in Cons::iter(mu, args) {
-                    let string = Cons::car(mu, cons);
-
-                    match string.type_of() {
-                        Type::Vector if Vector::type_of(mu, string) == Type::Char => {
-                            let str = Vector::as_string(mu, string);
-                            argv.push(str)
-                        }
-                        _ => return Err(Exception::new(Condition::Type, "system", string)),
-                    }
-                }
-
-                let status = std::process::Command::new(Vector::as_string(mu, command))
-                    .args(argv)
-                    .status();
-
-                match status {
-                    Err(_) => return Err(Exception::new(Condition::Open, "spawn", command)),
-                    Ok(exit_status) => Fixnum::as_tag(exit_status.code().unwrap() as i64),
-                }
-            }
-
-            Err(e) => return Err(e),
-        };
-
-        Ok(())
-    }
-
-    fn std_exit(_: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let rc = fp.argv[0];
-
-        match rc.type_of() {
-            Type::Fixnum => std::process::exit(Fixnum::as_i64(rc) as i32),
-            _ => Err(Exception::new(Condition::Type, "exit", rc)),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {}
